src/server/requirements.txt
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
beanie==1.17.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.0
fastapi==0.89.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
lazy-model==0.0.5
lxml==4.9.2
motor==3.1.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
passlib==1.7.4
pluggy==1.0.0
pydantic==1.10.4
pymongo==4.3.3
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.22.0
toml==0.10.2
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
